moved laser server stuff to more appropriately named directory
