+++ lightlab/equipment/lab_instruments/EMCORE_microITLA_server/EMCORE_microITLA_LS_server.py
@@ -17,7 +17,7 @@ import zmq
 # NEC
 try:
     from serial.tools.list_ports import comports
-    from itla_msa_modules.emcore_app_pyserial import emcore_app as app
+    from emcore_app_pyserial import emcore_app as app
     import serial
 except:
    raise
