+++ lightlab/equipment/lab_instruments/EMCORE_microITLA_server/install_service.sh
